chip/stm32/usb_hid_keyboard: Fix set_keyboard_report race
Previously, 2 consecutive calls to set_keyboard_report could lead
to the first one to be ignored, if the data did not have enough
time to be transmitted on the endpoint.

We fix that by setting an atomic variable to 0 when the endpoint
is busy, making sure the previous data is sent before erasing the
buffer.

RANCH=none
BUG=chrome-os-partner:59083
TEST=make buildall -j
TEST=make BOARD=hammer -j && util/flash_ec --board=hammer

Change-Id: I9d14541b8b05017c1d5051b9a315db381a89dcea