Update patch set 1
Patch Set 1:

(1 comment)

this seems like a bug in llvm.  i'd understand the warning if we hadn't forced the alignment to 4, but we did, so this code shouldn't be unaligned.  please file a bug upstream to deal with it and reference it in the code that the cast is a workaround.

Patch-set: 1