reef: Enable tablet mode
GPIO_TABLET_MODE is renamed to *_L to reflect its active-low configuration.
Note that the schematics needs correction as well.

BUG=chrome-os-partner:58792
BRANCH=none
TEST=Flip the lid over to the tablet zone and back to the laptop zone and
verified tablet_mode flips as expected. Flipped the lid to 360 degree and
verified GPIO_TABLET_MODE_L turns on (L) and EN_P3300_TRACKPAD_ODL turns
off (H).

On a touchview enabled system, verified a chromium window went to full
screen when the lid was flipped to the back and came back to a normal
window when the lid was flipped to the front.

Keyboard and trackpad are also disabled in tablet mode and re-enabled in
laptop mode.

Change-Id: I48c04bd576f457a899dfdf9b4718d73b59419cbe
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>