Reef: Disable keyboard and trackpad in tablet mode
This change disables the keyboard and the trackpad when the lid goes
beyond 180 degree.

It removes the tablet mode interrupt, which doesn't have any specific
task to do because tablet mode is handled by the lid angle detection.

BUG=chrome-os-partner:58792
BRANCH=none
TEST=Keyboard and trackpad are disabled when the lid goes beyond 180
and re-enabled when it's smaller than 180. Keyboard and trackpad are
disabled when the lid goes to 360 degree and the system doesn't wake
up by a keypress.

Change-Id: I48c04bd576f457a899dfdf9b4718d73b59419cbe
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>