Cr50: Prevent rebooting when unlocking the console
There were several subtle problems causing this. When the console
is unlocked, the function nvmem_wipe_or_reboot() is called. This
holds the EC in reset, clears nvmem, resets the TPM task, then
releases the EC.

First, putting the EC in reset invoked the sys_rst_asserted()
interrupt handler, triggering extra (and early) calls to
tpm_reset().

Second, the call to tpm_reset() caused the current (HOOKS) task
to wait for the operation to finish, but it didn't wait long
enough. When the task_wake_event() returned, a timeout was
indicated in addition to the completion event.

Third, because we checked for the timeout first, we reported an
error even though tpm_reset() completed successfully, although
slower than we expected.

Fourth, in some cases, recreating the endorsement credentials
took long enough that the watchdog fired before it finished.

This CL addresses all of these cases, plus adds a debugging
message in the (hopefully unlikely) event that any of these
problems resurfaces.

BUG=chrome-os-partner:59902
BRANCH=none
TEST=make buildall, manual tests

From the Cr50 console, run the "lock on" and "lock off" commands.
Try it both with and without the battery present. Observe that
the Cr50 no longer reboots just because the console unlocks.

Change-Id: I65a342502718acc5b9bda8c6f28dcd27e8f027f7
Signed-off-by: Bill Richardson <wfrichar@chromium.org>