Update patch set 1
Patch Set 1: Code-Review-1

I am not sure this brings much value - checking just 4 bytes to decide if the image is valid or not - the image could be invalid even if the first four bytes are set to the expected value.

Also, doesn't the code report 'error' in the image version field in case some header fields are not set to the expected values?

Patch-set: 1