Update patch set 1
Patch Set 1:

> Well, I figured it might be handy for distinguishing between prod
 > and dev-signed images, too. If we decide to invalidate an image for
 > incompatibility or bad fuse values or something, instead of just
 > ignoring it, this would help us know what happened to it.

I totally for adding prod vs dev bit to the version string, still, but XX does not convey this information.

How about changing the code to corrupt a field which is used to determine if the image is good or not, so when the RW is corrupted because RO started the other RW as a fallback, it will start showing as Error.

In any case it would be better to discuss the options on a tracker entry.

Patch-set: 1
Label: Code-Review=0