Update patch set 4
Patch Set 4:

If the battery is not providing voltage and CHG_EN is not set, VBATT should be 0V, right? So, when we come up in this state (VBATT = 0V, battery not responding to i2c?, yet bat present pin is asserted) can't we just take the necessary steps to wake the battery from shipmode?

Why do we need to set CHG_EN at hook_init time?

Patch-set: 4
Label: Code-Review=0