Update patch set 3
Patch Set 3:

(3 comments)

> Patch Set 2:
> 
> (1 comment)
> 
> My brain starts to be blurry, but what about this logic instead:
> 
> Given the reports are small, maybe we can do the copy to the usb_ram in interrupt context to simplify a little bit.
> 
> static int valid_index = -1;
> static int curr_index;
> static uint64_t last_report[2];
> 
> void set_keyboard_report(uint64_t rpt)
> {
>         int ongoing;
> 
>         curr_index = !curr_index;
>         last_report[curr_index] = rpt;
>         valid_index = curr_index;
>         
>         ongoing = (STM32_USB_EP(USB_EP_HID_KEYBOARD) & EP_TX_MASK) == EP_TX_VALID;
>         /* it will pick the buffer in IT context */
>         if (ongoing)
>              return;
>          
>         /* nothing on-going: prepare and start a new transfer */
>         memcpy_to_usbram((void *) usb_sram_addr(hid_ep_buf),
> &rpt, sizeof(rpt));
>         valid_index = -1;
>         STM32_TOGGLE_EP(USB_EP_HID_KEYBOARD, EP_TX_MASK, EP_TX_VALID, 0);
> }
>  
> static void hid_keyboard_tx(void)
> {
>     if (valid_index > 0) {
>         /* we have another report ready */
>         memcpy_to_usbram((void *) usb_sram_addr(hid_ep_buf), last_report+valid_index, sizeof(uint64_t));
>         valid_index = -1;
>         STM32_TOGGLE_EP(USB_EP_HID_KEYBOARD, EP_TX_MASK, EP_TX_VALID, 0);
>     }
> 
>     hid_tx(USB_EP_HID_KEYBOARD);
> }

I ended up doing something fairly similar, expect that I just update the address in the handler (since the buffer is small, we can also just use twice the amount of USB memory).

One problem with the code above is if you have A+B keys pressed within a single USB period, you'll get A in the first report, and A+B in the second (so there'll be a 1 period delay). I tried to compensate for that in my code my updating the address, and try to send A+B immediately, but then that might race with the hardware, so I have no way to tell if A or A+B was sent, so I need to queue A+B anyway (hence the duplicate report, which is actually harmless and ignored by HID).

Patch-set: 3
Label: Code-Review=0