Update patch set 2
Patch Set 2:

(1 comment)

I thought about it a bit, and I couldn't think of a way for that to happen because of the way tasks are prioritized, so I decided to keep the Cl small. OTOH, it does move all the clear-TPM stuff into the TPM task itself...  Yeah, okay, let's do that.

Patch-set: 2