Update patch set 3
Patch Set 3:

I don't see it as logically sistinct - that still would be a patch preventing reset during unlocking the console.

And it would be better, not leaving a window of opportunity for screwups this patch does.

Could be as simple as sending a different signal to the tpm task: do nvram wipeout before tpm reset.  I might me missing something, but I don't see where it is complicated.

Patch-set: 3