Update patch set 1
Patch Set 1:

I'm trying to plan ahead.

tpm_reset() needs to be non-blocking because it's invoked by SYS_RST_L and it can't block in an interrupt handler. But if we wanted to have a "tpm_reset" console command (which we might for several reasons), then it should block so we can observe the process.

tpm_wipe_and_reset() may or may not need to be blocking, but because it's currently called as part of unlocking the console in CCD, we want to reboot on failure to ensure that the console gets locked again. So for that, we can have it turn the AP off. But, if the AP wants to wipe and reset (say as part of a recovery image), then we don't want it to turn the AP off, because it might instead want to retry the wipe on failure without rebooting.

Patch-set: 1