Improve readability for file-scoped variables
File-scoped variables in common/tpm_registers.c must be handled
specially to avoid unexpected behavior when calling tpm_reset().
Enhance the comments and add a couple of macros to help explain
it better.

BUG=none
BRANCH=none
TEST=make buildall; Run on Reef

This is just commenting and code cleanup. No changes in behavior
are expected; none are observed.

Change-Id: If70e56d00642a11df7b5ceb5d5d32c485236f7a8
Signed-off-by: Bill Richardson <wfrichar@chromium.org>