nvmem: Modified nv_commit to only write blocks that will change
Previously, the entire NvMem partition was being erased and written
each time a call to nvmem_commit() was made. This process was
consuming ~80 - 110 msec (depending on whether the new partiion was in
the same half of flash as the RW code that's being executed.

This CL modifies nv_commit() so that each block that's about to be
written to flash is first compared with the existing contents. If
there is no change, then the erase/write steps are skipped. For TPM
operation, typically 2 blocks will need to be modified. For this case,
nv_commit() time is reduced to ~31 - 40 msec.

BRANCH=none
BUG=chrome-os-partner:59873
TEST=manual
Rebooted Reef and verified that "tpm_manufactured: manufactured" was
on the console output and there were no errors related to NvMem not
being initialized properly. Removed battery power and repeated the
same test. Also, verfied that the partiiion selection of 0|1 matched
the expected value based on the last partition written. (note had
debug code to print the partition being written and timestamp that was
removed prior to uploading the CL).

Change-Id: Ia5f3aea493fc5eba3cd55112dfc2ad65d6ae3616
Signed-off-by: Scott <scollyer@chromium.org>