pd_task: Add CC debounce time in DRP auto toggle
When TCPC is put into low power mode and type-c cable
is plugged in it returns to normal mode and then pd_task
reads CC values to determine the type of device connected.
In case of Analogix it is observed that CC values read are
invalid and needs around 100ms debounce time.

BUG=chrome-os-partner:60021
BRANCH=none
TEST=On Reef let analogix enter low power mode, and then plug in
hoho/apple hdmi dongle and verified if it is detected. Verified
multiple iterations.

Change-Id: Ibf4d49e64b07d18b6d1402788b7b549fce80ac3c
Signed-off-by: Divya Sasidharan <divya.s.sasidharan@intel.com>