Update patch set 1
Patch Set 1:

(4 comments)

The API and use of this feature need to be better described.

I take it it will allow for only ascii strings for key and value, save and and write env are different things. 

A comment block in the head of the file describing this would go a long way.

Patch-set: 1
Label: Code-Review=0