motion_lid: Add more reliability measurements.
Previously in motion lid, we only considered the lid angle as unreliable
when the hinge is too closely aligned with the direction of gravity.
However, there are other cases where the lid angle can be unreliable.
For example, when the device is being shaken and is under acceleration
that's not solely due to gravity.

This commit adds some more checks for a reliable lid angle measurement.

 - Checking if the device is significant motion by checking the
   deviation of the magnitudes of the base and lid vectors.

 - Making sure that the calculated angles agree with the current state
   of the lid switch.

BUG=chrome-os-partner:59480
BUG=chrome-os-partner:59203
BUG=chrome-os-partner:58518
BRANCH=gru,cyan,glados,oak
TEST=Flash caroline; use ectool motionsense lid_angle and monitor the
instantaneous lid angle.  Verify that unreliable is reported for cases
where the device is under significant motion.

Change-Id: I4bd9e818e617b056364cce2e46385e743a7522d4
Signed-off-by: Aseda Aboagye <aaboagye@google.com>