Update patch set 1
Patch Set 1: -Code-Review

> Patch Set 1: Code-Review+2

Actually, I'm not very sure about this. One case I'm concerned about is a lid is not completely flipped but stays at tablet zone (180 ~ 360 range). When the device is in that state, we should not enable input devices. With this change, if EC somehow boots up while the system is in S3, it'll enable the input devices, causing the system to wake up because the keyboard is facing down on the desk.

Patch-set: 1