Update patch set 1
Patch Set 1:

> > Patch Set 1: Code-Review+2
 > 
 > Actually, I'm not very sure about this. One case I'm concerned
 > about is a lid is not completely flipped but stays at tablet zone
 > (180 ~ 360 range). When the device is in that state, we should not
 > enable input devices. With this change, if EC somehow boots up
 > while the system is in S3, it'll enable the input devices, causing
 > the system to wake up because the keyboard is facing down on the
 > desk.

"EC boots up while the system is in S3" is not a valid situation. If the EC is booting the that means the AP is not powered up yet. The only way you could be in that situation is boot then goto s3. Then sysjump while sleeping. That's an artificial situation since it's impossible to do w/o a servo.

Patch-set: 1