Update patch set 2
Patch Set 2:

I know where our bad measurement comes from: as you mention, when the device is dropped, the results are off and can trigger the measurement out.

In chrome, we check if the vector length is too different:

bool IsAngleBetweenAccelerometerReadingsStable(
    const chromeos::AccelerometerUpdate& update) {
  return std::abs(
             ui::ConvertAccelerometerReadingToVector3dF(
                 update.get(chromeos::ACCELEROMETER_SOURCE_ATTACHED_KEYBOARD))
                 .Length() -
             ui::ConvertAccelerometerReadingToVector3dF(
                 update.get(chromeos::ACCELEROMETER_SOURCE_SCREEN))
                 .Length()) <= kNoisyMagnitudeDeviation;
}


We should do the same in motion_lid: what happen is one accel detected the drop: (vector length ~ 0), the other did not (sample later) and the angle was then random.

We can compare the length and also check the length: if != from ~10ms, the lid/base is in movemnent, we are not measure the gravity field.

Patch-set: 2
Label: Code-Review=0