stm32: Remove useless code from usb_power.c
The function usb_power_map_error() is only called from a single
location, with a fixed value passed in, and the return value, which
is the only reason for the function, is ignored.

The whole thing is useless, so remove the function and call.  Since the
sizes of all platforms are unchanged before and after removing this, I
suspect it was getting optimized out anyway.

Fixes coverity warning 157563: Useless call

BUG=None
BRANCH=None
TEST=build tested only.

Change-Id: Ie6572d2f5579819f06a53a567e1a93aabca217db
Signed-off-by: Martin Roth <martinroth@chromium.org>