Update patch set 6
Patch Set 6: Verified+1 Commit-Queue+1

> what happens if we haven't finished resurrecting the battery at this point 

Typically we will begin waking the battery immediately on RO power-up, and we must know we have a battery with sufficient charge before we allow the AP to boot. Yes, there are some battery modes where VBAT is not provided, yet when probed, the battery will respond mostly that everything is A-OK (CONFIG_BATTERY_REVIVE_DISCONNECT). We do need to make sure we kick the battery out of that state before booting the AP.

In the bigger picture, we do need to make sure that the system can subsist on 5V / 500mA for a brief period, even if a battery is present. That probably means inhibiting AP power-on for extremely low state of charge, regardless of external charger ability.

Patch-set: 6
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>