Update patch set 1
Patch Set 1:

The point of including an implementation's header file first (in this case the inclusion of "trng.h" in "trng.c") is to ensure that the header file is self contained. It it's not self contained, the compile will fail.

If the header isn't self contained, users will need to figure out what headers need to be included before including it. That doesn't make much sense.

Patch-set: 1