g: remove the impact of deep sleep resumes from the rolling reboot count
An image is marked good when there is a successful TPM read. Until the
image is marked good RO increments the retry counter on every reset to
try and catch a rolling reboot and fall back to the old image. This
includes deep sleep resets. If the TPM is not active cr50 can enter deep
sleep and resume for various reasons. We should consider these deep
sleep resumes ok and should not add to the retry counter. This change
subtracts one from the reset count when the system sees it just resumed
from deep sleep while the image is still being verified.

BUG=chrome-os-partner:60449
BRANCH=none
TEST=manual
	check the reset counter

	turn off the AP

	wait for cr50 to enter deep sleep

	plug in suzyq

	check it resumes from deep sleep and that the reset counter
	still has the same value

Change-Id: Ie8490c29636403b409b2a3f0912a5b312d23bc24
Signed-off-by: Mary Ruthven <mruthven@chromium.org>