g: remove the impact of deep sleep resumes from the rolling reboot count
The reset counter is used to detect a rolling reboot. Deep sleep resets 
should not contribute to that count. This change subtracts one from the
reset count when the system sees that it just resumed from deep sleep.

We still need a way to verify the TPM still works on a new image before
marking the image good.

BUG=chrome-os-partner:60449
BRANCH=none
TEST=manual
	check the reset counter

	turn off the AP

	wait for cr50 to enter deep sleep

	plug in suzyq

	check it resumes from deep sleep and that the reset counter
	still has the same value

Change-Id: Ie8490c29636403b409b2a3f0912a5b312d23bc24
Signed-off-by: Mary Ruthven <mruthven@chromium.org>