Update patch set 3
Patch Set 3:

(3 comments)

I'm thinking of adding back the gyro in for reliability for S0 only. It was definitely helpful. I don't think that we would need it in S3, since the new code to handle when the lid is closed. Going to S3 in tablet mode might be interesting though... could add some clamping based on the lid state and power state.

Patch-set: 3