cr50: keep wp and console state through deep sleep
After every reboot, we were resetting the write protect and console lock
states back to default. With this change the wp and lock states will be
preserved through deep sleep and reset on any other type of reboot. The
states are also cleared if the system detects a rollback.

If there was a rollback then the state wont be able to be saved until
the inactive rw image is updated or until the header is corrupted.

BUG=chrome-os-partner:58961
BRANCH=none
TEST=manual
	On prod/dev images verify that the default wp and console lock
	states are still correct.

	change the lock and write protect states from the default and
	verify they are preserved through deep sleep.

	reboot cr50 and make sure that they are reset.

	reboot cr50 4 more times.

	change the wp and lock states from the default setting.

	poweroff the ap

	wait until cr50 goes to sleep

	On resume from deep sleep check that the cr50 firmware rolled
	back and that the lock/wp states are reset to default

Change-Id: Ie7e62cb0b2eda49b04a592ee1d0903e83246b045
Signed-off-by: Mary Ruthven <mruthven@chromium.org>