kevin: Force keyscanning disable in tablet mode.
With the recent changes to motion lid, now if the angle is near 360
degrees, but we calculate it to be some really small angle, we mark the
angle as "unreliable".  So it's possibly that the lid_angle_update()
function doesn't get to disable keyscanning.  This commit adds a board
level callback to disable keyscanning when in tablet mode, just in case
it wasn't already.

BUG=None
BRANCH=gru
TEST=flash kevin; run `powerd_dbus_suspend --delay=5'; fold kevin back
to 360 degrees and set on table within 5 seconds.  Wait for suspend,
verify that keypress does not wake up device.

Change-Id: Ie9290437d2d8a8c2550088cc06c0fae59417a097
Signed-off-by: Aseda Aboagye <aaboagye@google.com>