Electro: Enable accels in S3
This patch enables accels in S3. Accels are required to calculate a
lid angle. EC enables/disables keyboard based on lid angles. EC
needs to be able to control it because the kernel is in sleep state
in S3.

BUG=chrome-os-partner:58792
BRANCH=reef
TEST=1. Put system into S3. Flip the lid to tablet zone. Hit keyboard to
verify the system doesn't wake up.
2. Put system into S3 in tablet mode. Flip the lid back to laptop zone.
Hit keyboard to verify the system wakes up.

Change-Id: I13c69a47da2c6521cd0c03c66cf061deb3f4fabd