Update patch set 1
Patch Set 1:

Booting it on some TPM2 device makes sense. Actually, booting and going through corp-enrollment to make sure RSA continues to work as expected (I don't see why it shouldn't, but it wouldn't be the first time a smoke test caught an issue with a seemingly unrelated code).

But we don't use ECC in the OS atm, so there are no ECC-specific tests to run.

Patch-set: 1