servo_v4: allow small RO with larger RW
This adds custom-ro_objs and SECTION_IS_RW checks
to servo v4. The intent is to resize RO and RW partitions
to allow more space for USBPD code. This change does not yet
redo flash layout, and does not automatically jump to RW
on a timeout, or use any other scheme to improve useability.
This assumes that RO config options are a subset of RW.

BRANCH=none
BUG=chrome-os-partner:61170
TEST=both partitions boot

Change-Id: I823ff08448e7320e6872db4fde3d17664a67cae5