lid_angle: wake up system when lid crosses 180 boundary
REFERENCE_ONLY & NOT_FOR_REVIEW

This change makes EC wake up the system when the lid goes from front
to back or back to front.

There is a discussion about why EC should not control trackpad. So, this
CL assumes EC does not control trackpad.

Without this change, the trackpad will stay on (thus can wake up the system
unintentionally) when the system enters S3 in laptop mode then
transitions to tablet mode, or vice versa: the trackpad will stay off
(thus cannot wake up the system) when the system enters S3 in tablet mode
then transitions to laptop mode.

This change gives the OS an opportunity to enable or disable when the lid
mode changes so that the trackpad can be used as a wake-up source properly.

This also cleans up the code. In terms of how peripherals are disabled or
enabled, there is no change in the functionality except:

lid_angle_peripheral_enable won't be called if there is no lid angle mode
(back <-> front) transition.

BUG=chrome-os-partner:61218
BRANCH=none
TEST=Put Electro into S3. Wake it up by flipping the lid. Verify small
angles do not wake up the system.

Change-Id: I8b08d5b11c47730584d33e57a920ec0933052fd8
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>