motion_sensor: fallback to laptop when failed to init required sensors
Since we enabled tablet mode in power_manager, we aren't able to use
keyboard and trackpad on some under develop devices whose lid accel
sensor is broken or not connected.

We cannot set the default mode to laptop for case of booting in tablet
mode. But we can fallback to laptop mode when something wrong with the
required sensors.

BUG=chrome-os-partner:61141
TEST=flash my kevin(with broken lid accel sensor), check keyboard.

Change-Id: Ic67c69e7c5b81b95b1c496b11fd2e54c8cf8c515
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>